Move all old pages to /widget
