+++ src/components/auth/common/polkadot-connect/PolkadotConnectWalletContent.tsx
@@ -3,7 +3,7 @@ import Button from '@/components/Button'
 import PopOver from '@/components/floating/PopOver'
 import MenuList, { MenuListProps } from '@/components/MenuList'
 import useIsInIframe from '@/hooks/useIsInIframe'
-import { ACCOUNT_SECRET_KEY_URL_PARAMS } from '@/pages/account'
+import { ACCOUNT_SECRET_KEY_URL_PARAMS } from '@/pages/widget/account'
 import { useSendEvent } from '@/stores/analytics'
 import { useMyAccount } from '@/stores/my-account'
 import { cx } from '@/utils/class-names'
